backend/requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
pdfplumber
python-docx
google-generativeai
pydantic
